+++ CHANGELOG.md
@@ -16,6 +16,7 @@ We do not (yet) follow semantic versioning.
 - introduce dg::assign and dg::construct that replace the dg::blas1::transfer functions and take an additional optional parameter pack
 
 ### Changed
+- Interface for Nprofile and ZonalFlow
 - Rename BinaryFunctors to CylindricalFunctors
 - Rewrite dg::ClonePtr as a pointer class
 - Change grids method interface of Multigrid to grid(unsigned)